src/backend/requirements.txt
psycopg2-binary
google-genai
tqdm
pandas
numpy
pathlib
toolbox-langchain
langgraph 
langchain-google-vertexai
flask
flask-cors
fastapi
uvicorn
gunicorn
google-cloud-storage
